requirements.txt
requests>=2.31.0
Pillow>=10.0.0
pdf2image>=1.16.0
dbus-python>=1.3.2
PyGObject>=3.42.0
